Split product validation snapshot projection
